Create README.md
This is a web application that implements PHP and MySQL in order to produce a final product which serves as a contact manger. Users can add a contact through the application as well as update and delete contacts. There is also a styleguide provided which includes the styles that were implemented in the application as well as other potential styles which could be.